Crate sui_proc_macros

Macros

checked_arithmetic

Attribute Macros

init_static_initializers

sim_test

The sim_test macro will invoke #[msim::test] if the simulator config var is enabled.

sui_test

The sui_test macro will invoke either #[msim::test] or #[tokio::test], depending on whether the simulator config var is enabled.

with_checked_arithmetic

Derive Macros

EnumVariantOrder

This proc macro generates a function order_to_variant_map which returns a map of the position of each variant to the name of the variant. It is intended to catch changes in enum order when backward compat is required.